html entity decoding in link list
The editor plugin that enables selecting an internal link gets
unreadable if ampersands or quotes are used in a page title, because
they are stored encoded.

For the link list, I believe this is not necessary. By decoding the
titles, the link list is much more usable by end users.