Added a logo for the Metro UI in Win8 & IE 10
The size of the image is the same as the image for the Ipad, but don't use the
same image, nor a full color image. Beter use a white monochrome version of the
logo, this will make it blend in with the Metro UI system tiles.

See http://kremalicious.com/add-your-web-site-to-the-windows-8-metro-ui/ for
more information.