Bugfix: do not add Spoon headers to the Response.
Spoon has already send the headers. Adding them to the Response object will
cause them to be send again. This caused all headers to be added twice and in
the event of a redirect caused the browser to halt because of duplicate Location
headers.

We should refactor out setting headers via Spoon, this is already being discussed in issue #402.