FIX: Form token usage in form builder.
When submitting a form with the form builder, this results in a
"something went wrong"; this is because the form generates a form token,
but since the form is built manually in the template, not taking
form_token into account, this results in a mismatch between the form
token in the session and that coming in from the form (that is to say;
none is coming in from the form).

Unfortunately, we cannot resolve to using {form:{$formName}} because
SpoonTemplate does not parse this correctly, so we built it manually
again in the template.

A better solution might be to fix SpoonTemplate so that it parses
{form:{$formName}} correctly, but I'd rather not touch external
dependencies and fork code in 1 change.

"accept-charset" when dealing with UTF8 was also broken, this changeset
fixes that too.

Closes #533