Authentication: do not allow God users to access uninstalled modules.
As a God user you are granted access to any module, even if it was not
installed. This caused database exceptions (missing tables) when
accessing the groups rights interface. This interface uses the
filesystem to list available modules and validated this with the
isAllowedX methods. This always returned true for God users.

Fixes #538