Mailmotor subscribe widget uses form token.
The Mailmotor widget for subscribing was not using FrontendForm to
create the subscription form.  This meant that the form token was not
generated and the receiving action (which does use form token) bailed
out.

This change set generates the form 'as it should', thus allowing
FrontendForm to generate the form token and the receiving 'subscribe'
action processing the form correctly.

I did not see an issue upon first sight, so there is no "closes..." part
in this commit message, if there was, I'm sorry ;-)