Optimise composer install
From the composer documentation: --optimize-autoloader (-o): Convert PSR-0 autoloading to classmap to get a faster autoloader. This is recommended especially for production, but can take a bit of time to run so it is currently not done by default.

I tested it and it doesn't seem to take a lot of time, so why not?