Fix bug with too much entries in locale/analyse
The BACKEND_MODULES_PATH is not a realpath, and is being used to
str_replace another path. This can't work, and therefor, some strings
are considered Non-existing. This commit adds realpath() function calls
for both the BACKEND_MODULES_PATH and the path of the module, so that
the str_replace can work correctly.