Avoid the 'The file "/private/nl/blog/index" does not exist' error when you have nog enough rights
When you want to access an action that exists but you have not enough rights to open it, the error action tried to create a Symfony File object with the url as path. 
This was done to get the extension. Since this is a url and not a path, the file could not be created causing a Symfony\Component\HttpFoundation\File\Exception\FileNotFoundException

This pull requests makes sure the correct error is shown ({$msgActionNotAllowed})