Trigger an event on users undelete action success
All other users actions have handy triggers on which you can hook in
other modules. The undelete action should have one too... That's the why
of this commit.
I'm deliberately not sending the BackendUser object along when I trigger
the event because it's a backend specific object and this could
theoretically be handled in frontend modules too. The user's id and
email should be enough in that case.