Fetch the SITE_DOMAIN from the request
This was done from the parameters, to make sure it works on CLI and in
browsers. By falling back to the headquarter domain, this is no issue
anymore.

Also: don't add the site.domain parameter to the parameters.yml file
anymore since it isn't used and isn't always relevant anymore"