Let prepare_for_reinstall also clean the database.
When developing and testing core functionality in Fork CMS, like
installers, multisite, etc... the install process needs to be run and
rerun a lot of times, manually cleaning the database is cumbersome, so
we've added the cleaning of the database with it.  We issue a warning
that this will happen, to notify people, maybe this should go in the
future?