Remove general settings in .editorconfig.
Upon further discussion, setting stuff like indents, line endings,
encoding, etc... for every file might not be such a good idea:  We might
have fixture files in a project that uses different encoding (and maybe
code, tests for instance, that depend on it!),  it might act
wrongly when code files in a non-configured language are added, and so
on.
Therefor, we remove the settings for all files entirely.
And add some likely languages like Python and Ruby to the mix.