Allow a site_id to be added to a setting.
Note: I also placed the creation of the site_title settings in the
multisite module. This is done becasue the getSites() function depends
on the sites table. This one isn't created yet in the CoreInstaller.

Adding a module setting with the overwrite option in a moduleInstaller
has also been changed.