Make Page and contentBlocks copy multisite.
I refactored both the page and contentblocks copying to an extra class.
This way, we have a nice (testable) class with decent dependency
injection.

The from_site and to_site parameters are added to the copy action. If
they aren't provided, we copy for the current site, to preserve
backwards compatibility.