Avoid unnecessary queries.
This commit lowers the amount of queries on the login page from 104 to 3
(with all base Fork modules installed).

Even for not loggedin users, it tried to fetch the allowed modules every
time an isAllowedAction or isAllowedModule was executed. This stayed
empty, so the same database call was done over and over again.