Bugfix right aligned tabs
In previous versions of Fork, the less important tabs were aligned on to the right side. The body element of the page was in 3.6.6: `body#pages.pagesAddEdit`.
This is now (apparently) divided into: `body#pages.pagesAdd`(if you add a page) and `body#pages.pagesEdit` (if you edit an existing page).

So the CSS rules needed to change to these two classes instead of one.

This commit fixes #766