Calculate default values for the database data.
Since we're not working with entities, we can't set default data in a
normal way. Using an event listener on the formBuilder, we can
manipulate the data right before it gets inserted in the form.

I haven't decided yet if we should use an object containing all
installation data instead off the parameterBag in the session.