Check if track cookie is not empty.
Due to the way SpoonCookie works, we get a lot of exceptions that the
value of the track cookie could not be retrieved.  By adding an extra
check on emptiness of the cookie we work around it.
This is not ideal, but it reduces the SpoonExceptions drastically.
Fixing it in Spoon would mean we need to throw out the exception
throwing, altering the behaviour of SpoonCookie unexpectedly for code in
user land.

Closes #845