Make sure our cache gets cleared when installing a new module.
Our installed modules are cached in the container. This means we need to
remove the cache to be able to use these newly installed modules.

Using the cache:clear command does not work, since this fucks up
sessions and thus causes a lot of errors.

You can also not overwrite the installed modules parameter in the
container since the container is frozen.