Fix "Illegal offset type" error
The setReplyTo method creates an array with the e-mailaddress as key and value. The variable $email is however an array with an 'id' and 'value'. Because this is an array, the error "Illegal offset type" is thrown on line 62. It should be a string, so make sure to fetch the value from $email.