Add tearDown method to Authentication.
The class retains values in the test and pollutes one login attempt with
previously set credentials. The introduction of the tearDown function
allows us to "reset" the class in the case that any one process would
log in, out and in again with different users.

This probably wont be a common real world scenario.