Bump the minimum required PHP version to 5.4.
We'll start following the supported versions of PHP. You can find the
graph here: https://secure.php.net/supported-versions.php.

Since PHP5.3 hasn't received security updates in 9 months, we don't want
to be responsible for user still building websites on it.

Note that the tests will still be run on 5.3, but are allowed to fail.