PSR-6 compatible navigation caching
Introduced a PSR-6 compatible cache. Filesystem-based by default
but can easily be swapped for other adapters (or even other
implementations, as long as they are PSR-6)

Navigation is no longer written to a file which is then included,
it'll now just read from *cache* (which may be non-filesystem)
This resulted in being able to drop a lot of code.