Encode the external redirect URL for pages
When a redirect URL has a double quote in it (not the best practice, but it
happens), it gives unexcepting results with the cache (because the cache is
JSON and the URL will be surrounded by double quotes. Adding the encoding for
the external redirect URL fixes wrong redirect URLs in navigation. Only add
encoding for the URL segments, because otherwise the whole URL (including ://)
will be encoded.

For example:
Redirect URL: http://www.google.be/Quote"HelloWorld"
Becomes: http://www.google.be/Quote in navigation