Make sure SEO tabs work again.
Every twig template is rendered in a different scope, so include tags
didn't really work for this.

The Node Parser for it isn't really clean, but it is the only way I
could think of to not loose the scope of the current form.

It's also a temporary solution since we'll eventually move to Symfony
Forms.