Make our position fixed again.
-webkit-transform causes our position to be fixed on our screen, not in
our document
(http://stackoverflow.com/questions/15194313/webkit-css-transform3d-position-fixed-issue)

An issue that arises on some screens (yeah, really, only on some screens) is that
in safari, you get a strange flicker when hovering over the arrow. It's
only on some screens and in Safari, so this is a better scenario than
keeping the position on a strange location.

We're looking for a fix for the flicker too.