Render each widget one at a time.
Previously, we added all variables first and then rendered one template
with includes. This resulted in errors when multiple instances of the
same widget but with different content were shown on one page.

As another good side effect, this fixes the infinite loop when we have
widgets containing random data.