Only show the tags tab when the user has permission
If the user had permission to edit/add pages but no access to the tags module
an alert with the error module not allowed was shown.
This was because the ajax call to get the tags would fail.
Now we only show that tabs if the user has sufficient permissions.

fixes #1562