Binding on `data-role="fork-*"` instead of classes
Because of the possible confusion that can exist on which classes we bind our
core JS we made the decision to not use classes/id's as a point to bind JS to.
Instead we will use the `data-role`-attribute.

As a result classes can be overwriten in themes without any problems.

Because `data-role` is a generic attribute it is possible that other plugins
use this attribute to, we namespace the items we bind to with the `fork-`-namespace.
So all "our" elements that have JS bind to, or JS relies on the elements will
have a data-attribute in the form of `data-role="fork-*"`.

Extra data that is need in JS can be passed thru extra data-attributes.

TL;DR: If you are creating new themes make sure to leave the data-attributes!