Fix notifying admin on blog commment
The function array_walk sends 2 parameters to the callback, the actual value and the key/index. Because of this
the rawurlencode function is not so happy and throws a warning that the function only accepts 1 parameter. Using
array_map, which only send 1 parameter (the value) to the callback, will fix this issue.