Added another variable "$startFromHighestParent"
This allows you to have a navigation with multiple layers down. For
example: if you are currently on a page from the fourth level, and you
set this as true, you will get a navigation from the 2th, third and
fourth level. Otherwise if set to “false” you will only see the
navigation from this fourth level