Catch Doctrine's DBALException of the database connection fails
On step 4 of the installer, the database variables are guessed from the URL of the website.
Fork then tries to connect to this database to see if it worked. If it doesn't it catches a ConnectionException but since DoctrineBundle 1.6.8 it also throws a DBALException if it can't determine the server version. We now catch that exception as well and treat it like a ConnectionException.

For more information see https://github.com/doctrine/DoctrineBundle/issues/673