Implemented a consent dialog
This consent should enable developers to build GDPR compliant websites. In the
backend the developer can define privacy levels to which the user can agree.

There is a service ForkCMS\Privacy\ConsentDialog which can be used to retrieve
the preferences of the current user.

The data is also available in jsData for usage in JavaScript