Removed visitor id
This feature conflicts with GDPR, although we had a way to not set the cookie.
As GDPR states that the user should know he is tracked, we can't generate a way
of tracking per default.

If you need this feature you should build it yourself, you can look at the past
code on how it was done, but you should ask permission first.