Function for retrieving the "index"
It is somewhat wierd code, but if you read https://stackoverflow.com/questions/175739/built-in-way-in-javascript-to-check-if-a-string-is-a-valid-number it will make sense.

The given string will be split into chunks based on _ as we know
this is the separator used. After this we will loop the chunks

The chunk is converted into a number (+"string"-trick), with isNaN
we can validate if it is NOT a number.

If we found the number we return it, otherwise we fallback