Method to import profiles from an array
This method is heavily inspired on importCsv. In essence it is a
working version. As at the moment I think importCsv will fail because
of fields that can't be NULL, but the code does not handle it.

I didn't fix importCsv because it would require database-structure changes
and won't be backwards compatible.